fix(oauth device): move DeviceCodeEntry out of src/pages (unbreak Next build) + autofocus input
PR #2647 added DeviceCodeEntry.tsx and DeviceCodeEntry.browser.test.tsx under
src/pages/login/oauth/. Next.js treats everything under src/pages/** as a route
and collects it during the build, so:

  - the Next build (build-image) failed: Next collected
    DeviceCodeEntry.browser.test as a page —
    "Type 'typeof import(... DeviceCodeEntry.browser.test)' does not satisfy the
    constraint 'PagesPageConfig'".
  - typecheck failed: the browser test's two `.rerender` calls were made on the
    result of the render helper (a Promise) without awaiting — TS2339
    "Property 'rerender' does not exist on type 'Promise<RenderResult>'".

This unbreaks main's Next build (RED since #2647 merged) and resolves PR #2648's
inherited preview failure (#2648's own e2e test under src/tests/ is correctly
located and was never the problem).

Changes:
- Relocate the component + its co-located browser test out of src/pages/ into
  src/components/Login/, matching the established "browser tests live beside the
  component under src/components" convention (Next ignores src/components for
  page collection; repo has no pageExtensions filter, so the move IS the fix).
  Only the real page route device.tsx (+ pre-existing authorize.tsx) remains
  under src/pages/login/oauth/.
- Repoint device.tsx's import and the test's import to the new path; fix the
  test's relative import depth to test/component-setup.
- Await the render helper before `.rerender` so typecheck passes (no any/ignore).
- Autofocus the code input on mount (ref + one-shot useEffect) so the user can
  paste/type immediately without clicking; the effect runs once and does not
  re-run on the auto-submit/loading re-renders, so it can't fight the existing
  auto-submit-on-completeness or Enter-to-submit logic. Adds a test asserting
  the input is focused on mount.
